refactor: move loading specific articles to article service
RISDEV-4322
